Re-wrote nessus parser to be more flexible
